pyproject.toml
[tool.poetry]
name = "brownie-bot"
version = "0.1.0"
description = ""
authors = ["Beier <Beierlejc@gmail.com>"]
readme = "README.md"
packages = [{include = "brownie_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
py-cord = "^2.3.2"
selenium = "^4.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
